+++ lizard_fewsnorm/models.py
@@ -94,6 +94,7 @@ class Parameter(models.Model):
 
     class Meta:
         db_table = u'parameters'
+        #db_schema = 'nskv00_opdb'
         managed = False
 
     def __unicode__(self):
@@ -161,7 +162,8 @@ class Timesteps(models.Model):
     timestepkey = models.IntegerField(primary_key=True,
                                       db_column='timestepkey')
     id = models.CharField(unique=True, max_length=64)
-    description = models.CharField(max_length=64)
+    #description = models.CharField(max_length=64)  # on testdatabase
+    #label = models.CharField(max_length=64)  # on fewsnorm-dev
 
     class Meta:
         db_table = u'timesteps'
@@ -202,9 +204,9 @@ class Series(models.Model):
     timestep = models.ForeignKey(Timesteps,
                                  db_column='timestepkey',
                                  blank=True, null=True)
-    aggregationperiody = models.ForeignKey(AggregationPeriods,
-                                           db_column='aggregationperiodkey',
-                                           blank=True, null=True)
+    aggregationperiod = models.ForeignKey(AggregationPeriods,
+                                          db_column='aggregationperiodkey',
+                                          blank=True, null=True)
 
     class Meta:
         db_table = u'timeserieskeys'
@@ -250,6 +252,11 @@ WHERE (locationkey, parameterkey) IN %s" % (keys,)).using(db_name)
 
 
 class Event(composite.CompositePKModel):
+    """
+    A single event.
+
+    Use the model Series to find out more.
+    """
     series = models.ForeignKey(Series,
                                primary_key=True,
                                db_column='serieskey')
@@ -263,11 +270,11 @@ class Event(composite.CompositePKModel):
         managed = False
 
     def __unicode__(self):
-        return u'%s %s %s %s' % (
-            self.id,
-            self.datetime,
-            self.scalarvalue,
-            self.flags)
+        return u'%s %s value=%s %s' % (
+            self.series,
+            self.timestamp,
+            self.value,
+            self.flag)
 
     @classmethod
     def filter_latest_before_deadline(cls, series_set, deadline):
@@ -344,6 +351,9 @@ class TimeseriesManualEditsHistory(models.Model):
 class ParameterCache(models.Model):
     ident = models.CharField(max_length=64)
 
+    class Meta:
+        ordering = ('ident', )
+
     def __unicode__(self):
         return '%s' % self.ident
 
@@ -355,6 +365,9 @@ class ParameterCache(models.Model):
 class ModuleCache(models.Model):
     ident = models.CharField(max_length=64)
 
+    class Meta:
+        ordering = ('ident', )
+
     def __unicode__(self):
         return u'%s' % self.ident
 
@@ -364,6 +377,9 @@ class ModuleCache(models.Model):
 
 class TimeStepCache(models.Model):
     ident = models.CharField(max_length=64)
+
+    class Meta:
+        ordering = ('ident', )
 
     def __unicode__(self):
         return u'%s' % self.ident
@@ -386,8 +402,16 @@ class GeoLocationCache(GeoObject):
         TimeStepCache, null=True, blank=True, through='TimeSeriesCache')
     objects = models.GeoManager()
 
+    class Meta:
+        ordering = ('ident', 'name')
+
     def __unicode__(self):
-        return '%s %s ' % (self.fews_norm_source, self.ident)
+        try:
+            return '%s (%s)' % (self.ident, self.fews_norm_source)
+        except:
+            # You can use GeoLocationCache as a memory object without
+            # defining fews_norm_source.
+            return '%s' % self.ident
 
     def api_url(self):
         return reverse('lizard_fewsnorm_api_location_detail',
@@ -548,7 +572,7 @@ class FewsNormSource(models.Model):
             time_series_cache.save()
 
     def __unicode__(self):
-        return '%s (%s)' % (self.name, self.database_name)
+        return '%s' % (self.name)
 
     def o(self, model_object):
         """